Move get_api_key to TrentoWeb.Auth.ApiKey module
