+++ lib/trento_web/controllers/v1/installation_controller.ex
@@ -4,6 +4,8 @@ defmodule TrentoWeb.V1.InstallationController do
 
   alias OpenApiSpex.Schema
 
+  alias TrentoWeb.Auth.ApiKey
+
   operation :get_api_key,
     summary: "Retrieve API Key",
     tags: ["Platform"],
@@ -24,7 +26,7 @@ defmodule TrentoWeb.V1.InstallationController do
 
   @spec get_api_key(Plug.Conn.t(), any) :: Plug.Conn.t()
   def get_api_key(conn, _) do
-    key = Trento.Settings.get_api_key()
+    key = ApiKey.get_api_key()
 
     render(conn, "api_key.json", api_key: key)
   end
