+++ test/trento_web/plugs/authenticate_apikey_plug_test.exs
@@ -2,6 +2,7 @@ defmodule TrentoWeb.Plugs.AuthenticateAPIKeyPlugTest do
   use TrentoWeb.ConnCase, async: true
   use Plug.Test
 
+  alias TrentoWeb.Auth.ApiKey
   alias TrentoWeb.Plugs.ApiAuthErrorHandler
   alias TrentoWeb.Plugs.AuthenticateAPIKeyPlug
 
@@ -17,7 +18,7 @@ defmodule TrentoWeb.Plugs.AuthenticateAPIKeyPlugTest do
   test "Accept an Authenticated request" do
     conn =
       conn(:get, "/foo")
-      |> put_req_header("x-trento-apikey", Trento.Settings.get_api_key())
+      |> put_req_header("x-trento-apikey", ApiKey.get_api_key())
       |> AuthenticateAPIKeyPlug.call(ApiAuthErrorHandler)
 
     assert %{assigns: %{api_key_authenticated: true}} = conn
